services/normalizer/requirements.txt
beautifulsoup4
readability-lxml
kafka-python
requests
python-dotenv
python-dateutil